+++ zetaclient/chains/solana/signer/withdraw.go
@@ -5,7 +5,7 @@ import (
 	"encoding/hex"
 
 	"cosmossdk.io/errors"
-	"github.com/gagliardetto/solana-go"
+	sol "github.com/gagliardetto/solana-go"
 	"github.com/near/borsh-go"
 	"github.com/rs/zerolog"
 
@@ -101,7 +101,7 @@ func (signer *Signer) createMsgWithdraw(
 }
 
 // createWithdrawInstruction wraps the withdraw 'msg' into a Solana instruction.
-func (signer *Signer) createWithdrawInstruction(msg contracts.MsgWithdraw) (*solana.GenericInstruction, error) {
+func (signer *Signer) createWithdrawInstruction(msg contracts.MsgWithdraw) (*sol.GenericInstruction, error) {
 	// create withdraw instruction with program call data
 	dataBytes, err := borsh.Serialize(contracts.WithdrawInstructionParams{
 		Discriminator: contracts.DiscriminatorWithdraw,
@@ -115,13 +115,13 @@ func (signer *Signer) createWithdrawInstruction(msg contracts.MsgWithdraw) (*sol
 		return nil, errors.Wrap(err, "cannot serialize withdraw instruction")
 	}
 
-	inst := &solana.GenericInstruction{
+	inst := &sol.GenericInstruction{
 		ProgID:    signer.gatewayID,
 		DataBytes: dataBytes,
-		AccountValues: []*solana.AccountMeta{
-			solana.Meta(signer.relayerKey.PublicKey()).WRITE().SIGNER(),
-			solana.Meta(signer.pda).WRITE(),
-			solana.Meta(msg.To()).WRITE(),
+		AccountValues: []*sol.AccountMeta{
+			sol.Meta(signer.relayerKey.PublicKey()).WRITE().SIGNER(),
+			sol.Meta(signer.pda).WRITE(),
+			sol.Meta(msg.To()).WRITE(),
 		},
 	}