+++ zetaclient/chains/ton/observer/observer_test.go
@@ -161,9 +161,9 @@ func (ts *testSuite) OnGetTransactionsSince(
 		Return(txs, err)
 }
 
-func (ts *testSuite) OnGetAllOutboundTrackerByChain(trackers []cc.OutboundTracker) *mock.Call {
+func (ts *testSuite) OnGetOutboundTrackers(trackers []cc.OutboundTracker) *mock.Call {
 	return ts.zetacore.
-		On("GetAllOutboundTrackerByChain", mock.Anything, ts.chain.ChainId, mock.Anything).
+		On("GetOutboundTrackers", mock.Anything, ts.chain.ChainId).
 		Return(trackers, nil)
 }