+++ changelog.md
@@ -27,6 +27,7 @@
 * [4200](https://github.com/zeta-chain/node/pull/4200) - remove `LastBlockHeight` state variable
 * [4174](https://github.com/zeta-chain/node/pull/4174) - add documentation for ZetaClient logging fields
 * [4210](https://github.com/zeta-chain/node/pull/4210) - skip writing config file to the filesystem when updating consensus timeout deltas
+* [4213](https://github.com/zeta-chain/node/pull/4213) - prepare the client interfaces of the observer-signers for dry mode
 
 ### Fixes